backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
qdrant-client==1.12.1
python-multipart==0.0.12
openai==1.52.2
tree-sitter==0.20.4
tree-sitter-languages==1.10.2
google-generativeai==0.8.3
anthropic==0.37.1
pytest==8.3.3
anthropic==0.37.1
